refactor: update Dice exception logic + test
