+++ src/test/java/edu/ntnu/idi/idatt/model/entities/DiceTest.java
@@ -1,10 +1,13 @@
-package edu.ntnu.idi.idatt;
+package edu.ntnu.idi.idatt.model.entities;
 
-import edu.ntnu.idi.idatt.model.entities.Dice;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertInstanceOf;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 /**
  * Test class for the Dice class.
